books

+++ appCourse/src/main/kotlin/com/github/hummel/mpp/course/entity/Book.kt
@@ -1,5 +1,5 @@
 package com.github.hummel.mpp.course.entity
 
 data class Book(
-	val id: Int, var title: String, val description: String, val author: String, val imgPath: String, val price: Int
+	val id: Int, var title: String, val description: String, val author: String, val imgPath: String, val price: Double
 )
